requirements.txt
cohere
weaviate-client
python-dotenv 